requirements.txt
asttokens==2.4.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
elastic-transport==8.15.1
elasticsearch==8.15.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.1
identify==2.6.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.0
wcwidth==0.2.13
